pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["senet_toolbox"]

[tool.hatch.build.targets.sdist]
exclude = ["notebooks", "tests"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = [
    "senet_toolbox/static_data/*.csv",  # Include CSV file in the package
]

[project]
name = "senet_toolbox"
version = "0.1.0"
description = "A Python package for Sen-ET"
authors = [
  { name="Marie Lund Larsen", email="mlla@dhigroup.com" },
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy==1.26.4",
    "numba",
    "pydms @ git+https://github.com/radosuav/pyDMS@v1.2",
    "pypro4sail @ git+https://github.com/hectornieto/pypro4sail@v1.2",
    "pytseb @ git+https://github.com/hectornieto/pyTSEB@v2.3",
    "meteo-utils @ git+https://github.com/hectornieto/meteo_utils@v2.1.1",
    "ipykernel",
    "openeo",
    "rioxarray",
    "pystac_client",
    "netcdf4",
    "ipyleaflet==0.19.1",
    "jupyter-leaflet==0.19.2",
    "scikit-image",
    "cdsapi",
    "cfgrib",
    "Py6S"
]

[project.optional-dependencies]
dev = [
  "build",
  "pytest",
  "ruff",
]

test= [
  "pytest",
]

notebooks = [
  "jupyter",
]

[project.urls]
"Homepage" = "https://github.com/DHI/Sen-ET-OpenEO-toolbox"
"Bug Tracker" = "https://github.com/DHI/Sen-ET-OpenEO-toolbox/issues"

[tool.ruff]
# ignore long lines
lint.ignore = ["E501"]
